requirements.txt
openai-whisper
sounddevice
numpy
deep-translator
pyttsx3
edge-tts
